First cut of node sdm
