+++ package.json
@@ -23,8 +23,10 @@
     "@atomist/spring-automation": "^0.6.2",
     "app-root-path": "^2.0.1",
     "artifactory-publisher": "^1.1.4",
+    "build": "^0.1.4",
     "copyfiles": "^1.2.0",
     "jssha": "^2.3.1",
+    "run": "^1.4.0",
     "strip-ansi": "^4.0.0",
     "xml2js": "^0.4.19"
   },
@@ -71,7 +73,7 @@
     "gql:gen": "atomist gql-gen \"./graphql/**/*.graphql\"",
     "lint": "npm-run-all lint:ts lint:gql",
     "lint:ts": "tslint --format verbose --project . --exclude \"{build,node_modules}/**\" \"**/*.ts\"",
-    "lint:gql": "prettier --list-different \"src/graphql/**/*.graphql\"",
+    "lint:gql": "prettier --list-different \"graphql/**/*.graphql\"",
     "lint:fix": "npm-run-all lint:ts:fix fmt:gql",
     "lint:ts:fix": "npm run lint:ts -- --fix",
     "start": "atomist start --no-install --no-compile",
