+++ src/atomist.config.ts
@@ -2,11 +2,10 @@ import { Configuration } from "@atomist/automation-client/configuration";
 import * as appRoot from "app-root-path";
 import { HelloWorld } from "./handlers/commands/HelloWorld";
 import { MachineAssembler } from "./sdm-support/MachineAssembler";
-import { applyHttpServicePhases } from "./software-delivery-machine/blueprint/phase/phaseManagement";
+import { applyHttpServicePhases } from "./software-delivery-machine/blueprint/phase/jvmPhaseManagement";
 import { affirmationEditor } from "./software-delivery-machine/commands/editors/affirmationEditor";
 import { breakBuildEditor, unbreakBuildEditor } from "./software-delivery-machine/commands/editors/breakBuild";
-import { springK8sSoftwareDeliveryMachine } from "./software-delivery-machine/springK8sSoftwareDeliveryMachine";
-import { springPCFSoftwareDeliveryMachine } from "./software-delivery-machine/springPCFSoftwareDeliveryMachine";
+import { cloudFoundrySoftwareDeliveryMachine } from "./software-delivery-machine/machine/cloudFoundrySoftwareDeliveryMachine";
 
 // tslint:disable-next-line:no-var-requires
 const pj = require(`${appRoot.path}/package.json`);
@@ -14,7 +13,7 @@ const pj = require(`${appRoot.path}/package.json`);
 const token = process.env.GITHUB_TOKEN;
 
 const assembled = new MachineAssembler(
-    springPCFSoftwareDeliveryMachine({useCheckstyle: false}),
+    cloudFoundrySoftwareDeliveryMachine({useCheckstyle: false}),
     // springK8sSoftwareDeliveryMachine({useCheckstyle: false}),
 );
 
@@ -23,7 +22,8 @@ export const configuration: Configuration = {
     version: pj.version,
     teamIds: [
         //  "T1JVCMVH7",
-        "T5964N9B7",
+        "T5964N9B7",    // spring-team
+        //  "T29E48P34",    // Atomist community
     ], // <-- run @atomist pwd in your slack team to obtain the team id
     commands: assembled.commandHandlers.concat([
         HelloWorld,
