+++ src/handlers/events/delivery/phases/npmPhases.ts
@@ -1,0 +1,8 @@
+
+import { Phases } from "../Phases";
+import { BaseContext, BuildContext, IndependentOfEnvironment, ScanContext } from "./gitHubContext";
+
+export const npmPhases = new Phases([
+    ScanContext,
+    BuildContext,
+]);
